Add basic oomox-gtk-theme derivation
For my current laptop configuration I maintain a custom GTK theme to match
i3 and vim color schemes (see KnairdA/dotfiles). As starting  point I want
to port this setup to NixOS and XMonad. The objective is to get to a point
where I can change accent colors in a central location (i.e. some Nix file)
and generate everything else automatically.

The included derivation builds and installs this custom oomox generated
theme in a quick and dirty fashion. This can  definitely be improved in
many ways, e.g. splitting the config into Nix params.
The result looks mostly correct - remaining differences are probably due
to the missing Icon theme.

Additionally I had to define $HOME as oomox seems to attempt to write
there in spite of the target directory definition.

Note that this derivation currently only builds on unstable as older
versions of sassc terminate with some kind of CSS error. Luckily this is
no problem for Nix.