Add Neovim-Qt as vim frontend
Default CLI editor is currently still Vim, mainly because cs-exact
doesn't work correctly in Neovim. Neovim-Qt is preferable to GVim as the
latter has ugly pixel glitches on my system (maybe related to
proprietary NVidia drivers? This was not a problem on the x220)

The neovim-qt derivation is built using a custom nvim derivation which
in turn uses the global vim configuration managed in /etc/nixos/pkgs/vim.