Improvise rofi task launcher
`tasker` opens a rofi selection of all scripts found in `~/.local/share/tasks`.
This is useful for launching e.g. project specific nix-shells, vim instances and so on.

Actual tasks are not included at this point as they are still quite rudimentary.

This vision is that this develops into a Nix and XMonad integrated
project manager that:

* Automatically clones the project repository if it is not already available
* Allows for (optional) central management of nix-shell derivations
* Allows for easy isolation of project-specific UI utilities
* Enables convenient launching of build and test scripts