Replace lorri with custom persistent-nix-shell
Lorri has shown itself to be rather unstable for my particular use case:

* It is unclear when derivations are actually rebuilt
* Polling did not seem consistent, i.e. I had to issue manual `ping_` calls
* I do not want to issue manual `lorri watch` calls to fix this
* It sporadically started building GC roots without any clear reason for doing so

In the end all I really want is a way to prevent nix-shells from being
garbage-collected. In this scenario using lorri is overkill.

`persistent-nix-shell` is a simple wrapper around `nix-shell` that adds
GC roots prior to entering the environment.